Fix broken cache dir path computation in CarrierWave::Uploader::Store#delete_cache_id
This fixes the broken spec "CarrierWave::Uploader#store! should delete the old cache_id" that was added in 7f9c35fa

This now breaks numerous other specs and support for uploaders with multiple versions, because now that the cache dir is successfully being deleted after the first version is stored, other versions in that cache are being prematurely deleted.