Removed a faulty test
The test that was removed was producing a false failure (or rather, test conditions were hiding a defect). Once the defect was fixed, the test began reporting a failure, since the test was written with expectations based on the defect.

If that doesn't make sense, the gist is that during testing setup, CarrierWave.root is explicitly set prior to loading in all the modules, such as CarrierWave::Uploader::Configuration; thus the test passed. When correctly replicating the order of startup, as in a Rails app, the defect was exposed and the test failed. So the test was not testing the right thing.

A replacement test was added (in an earlier commit, b4bc77270bd25a243e26513dac658a2f71c2d507) to properly test this scenario. I had left the failing test in place to demonstrate the failure, but since all is working correctly, the bad test has been removed.