FIX inconsistence in file API
Not sure if doing it right, but this commit intends to fix an inconsistency
in file access API.

We use an WorkFileUploader, inheriting from CarrierWave::Uploader::Base, mounted
in a mongo model.

On production, CarrierWave is configured to use fog as storage ; on
development env, we use the file storage.

This works as expected on dev env :

```ruby
model.work_file.file.filename # => 'document_name.ext'
```

`model.work_file.file` is a CarrierWave::SanitizedFile, in that env.

But on production env, it is a `CarrierWave::Storage::Fog::File`, which
has no `#filename` method.