Fixed bug in serializing to xml or json where both :only and :except are passed
There was a bug in determining if the uploader should be serialized
when both :only and :except were passed. :except should be taken into
account only if :only isn't passed

This bug was especially important with ActiveRecord's to_xml
implementation, which adds an additional :except => :type to the
options hash. This revealed the bug even when I was only passing :only