Adding Jruby to experimental test environments
I realize that Jruby is not officially supported (so the jruby test environments are separated into the 'allow_failures' area) but adding Jruby tests will give curious jruby users a good idea of how viable it would be to user Carrierwave with jruby. If there are not many errors, perhaps jruby users could step up and propose changes that might allow Carrierwave to support jruby. 

*NOTE* - This would *not* imply official support and would be for informational purposes only