Fix file delete being called twice on remove.
When a file is removed with something like `.remove_image = true` it's
seen as you replacing the file with a new file and still tries to delete
the "old file" even though the old file was just deleted. This is
mitigated in `CarrierWave::SanitizedFile#delete` by an `if exists?`
condition. But it's unnecessary and especially annoying when trying to
use a file storage API.