Remove `CarrierWave::MimeTypes` processor module
Since #1245 (*), CarrierWave::MimeTypes module is redundant as
`mime-types` gem is a runtime dependency now, and `SanitizedFile` gets
it's content_type using it.
The module was deprecated since 0.10 and warned developers when used it.

(*) I discovered this when tried to fix specs marked as pending. I
looked at commits history and PRs to understand what happened.
It started with #372 that added `CarrierWave::MimeTypes`  processor (Jun
2011), then #1216 and #1245 that made `mime-types` a runtime dependency
and deprecated `CarrierWave::MimeTypes` module (2013), and finally #1584
`CarrierWave::MagicMimeTypes` (2015)