Change #cache_storage to use #storage by default to avoid confusion
Most users won't expect that they must be configured separately,
like said in #2353. But we preserve ability to set cache_storage explicitly,
because there can be a case which gets benefit from it.
(e.g. For an application which utilizes multiple servers but does not
need caches to be persisted across requests, using :fog for storage and
:file for cache_storage results in smaller overhead)