Skip content type filters where possible.
Prior to this patch every content_type_* filter would parse magic headers,
even when the filter was not defined.

I have split the primary conditional check in to a guard clause that
simply returns unless the blacklist or whitelist returns truthy.

I've made the same changes to the extension filters for consistency.