Add "public" messages to errors
The message for CarrierWave::ProcessingError used to include
- the name of the underlying image processor (RMagick or
  MiniMagick) and
- the original error message produced by it.

The message is used for ActiveRecord validation, and it is often
shown for an end-user. This behavior has some downsides:
- The message may look cryptic and is confusing for end-users.
  A typical end-user has no idea what is RMagick, for example.
- The information the message carries may be valuable for an
  adversary. Especially, the image processor may not be careful
  enough to mask sensitive information in an error message it
  produces. e.g. MiniMagick may include stderr of ImageMagick,
  which may contain file paths and reveal underlying ImageMagick
  configurations, in an error message.

This change solves the problems by adding public_message
attributes, which does not contain such sensitive information, to
error types and exposing them in ActiveRecord validation.