Raise DownloadError when no content is returned
Have RemoteFile raise CarrierWave::DownloadError when the HTTP response
is of a type that has no content, according to the standards.
In those cases, response.body returns nil.

There are 8 response types that have no body, 4 of which we can get, so
just check whether the body was nil.

Fixes #2632