Move test for GovCloud or FIPS higher
The code can be simplified if we move the test for GovCloud or FIPS
outside of the if use_virtual_hosted_style block.  This makes it
explicit what it's trying to do - which is disable the use of
acceleration.

We won't emit a warning here.  The corresponding code going into
the fog-aws gem will emit a warning once that code is merged. We
don't need two warnings for the same thing.