Raise MiniMagick::Error if ImageMagick is missing delegates
Fixes #2732

Starting in Alpine Linux 3.19, installing the `imagemagick` package no
longer includes support for JPG, SVG, and other image formats.[1]

This means upgrading Alpine linux on a server can cause Carrierwave to
start failing with CarrierWave::ProcessingError when uploading images.

The root cause isn't obvious from this error.

However when ImageMagick is not installed, Carrierwave raises
MiniMagick::Error instead of CarrierWave::ProcessingError, and includes
the detailed error message.

This fix also raises MiniMagick::Error if ImageMagick is installed but
is missing the delegates required for the image format being
manipulated. This will make it easier to debug when uploads are failing.

[1] https://maxsmolens.org/posts/imagemagick-packaging-change-on-alpine-linux/